Drop 0.4 and 0.7 fixes
* Drop remaining 0.4 support code
* Deprecate remaining 0.4 Compat bindings
* Remove tests for 0.4 Compat
* Do not run deprecated tests on 0.6
* Recover one test that's incorrectly moved to deprecated (StringVector)
* Add compat for at-nospecialize with tests
* Add compat for `read(..., String)` with tests
