+++ src/deprecated.jl
@@ -143,6 +143,15 @@ end
 using .CompatCartesian
 export @ngenerate, @nsplat
 
+function primarytype(@nospecialize(t))
+    tn = t.name
+    if isdefined(tn, :primary)
+        return tn.primary
+    else
+        return tn.wrapper
+    end
+end
+
 export @functorize
 macro functorize(f)
     code = f === :scalarmax          ? :(Base.scalarmax) :
@@ -160,13 +169,22 @@ end
     Base.@deprecate_binding KERNEL Sys.KERNEL
     Base.@deprecate_binding UTF8String Core.String
     Base.@deprecate_binding ASCIIString Core.String
+    Base.@deprecate_binding unsafe_convert Base.unsafe_convert
+    Base.@deprecate_binding remote_do Base.remote_do
+    Base.@deprecate_binding Filesystem Base.Filesystem
+    Base.@deprecate_binding AsyncCondition Base.AsyncCondition
+    Base.@deprecate_binding promote_eltype_op Base.promote_eltype_op
+    @eval Base.@deprecate_binding $(Symbol("@irrational")) Base.$(Symbol("@irrational"))
+    @eval Base.@deprecate_binding $(Symbol("@blasfunc")) Base.LinAlg.BLAS.$(Symbol("@blasfunc"))
 else
     const KERNEL = Sys.KERNEL
     const UTF8String = Core.String
     const ASCIIString = Core.String
+    import Base.unsafe_convert
+    import Base.remote_do
+    import Base.Filesystem
+    import Base.AsyncCondition
+    import Base.promote_eltype_op
+    import Base.@irrational
+    import Base.LinAlg.BLAS.@blasfunc
 end
-
-# More things that could be removed in Compat.jl
-# - new_style_call_overload
-# - import Base.Filesystem
-# - import Base.LinAlg.BLAS.@blasfunc
